deep frier slice fix (#581)
* move SliceFoodEvent into a deltav file, move raising into Slice()

* import deltav event

---------
